homework_11-14/pyproject.toml
[tool.poetry]
name = "homework-2-11"
version = "0.1.0"
description = "REST API for contacts"
authors = ["Matajur <gavrylchenko.d@ukr.net>"]
license = "Apache License 2.0"
readme = "README.md"
packages = [{include = "homework_2_11"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
sqlalchemy = "^2.0.20"
psycopg2 = "^2.9.7"
pydantic = {extras = ["email"], version = "^2.3.0"}
pydantic-extra-types = "^2.1.0"
phonenumbers = "^8.13.20"
python-multipart = "^0.0.6"
alembic = "^1.12.0"
jinja2 = "^3.1.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
libgravatar = "^1.0.4"
fastapi-mail = "^1.4.1"
pydantic-settings = "^2.0.3"
fastapi-limiter = "^0.1.5"
redis = "<5.0.0"
cloudinary = "^1.36.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"
pytest = "^7.4.2"


[tool.poetry.group.test.dependencies]
httpx = "^0.25.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
